pyproject.toml
[project]
name = "pollardstrassen"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gmpy2>=2.2.2",
    "python-flint>=0.8.0",
]
